// src/utils/types.ts
export type AppSettings = {
  uppercase: boolean;
  lowercase: boolean;
  figure: boolean;
  characters: boolean;
  special: boolean;
  showModal: boolean;
};

export type Conditions = {
  uppercase: boolean;
  lowercase: boolean;
  figure: boolean;
  characters: boolean;
  special: boolean;
};

export type Form = {
  email: string;
  password: string;
};